requirements.txt
transformers >= 4.35.3
accelerate >= 0.26.0
anthropic
ultralytics
torchvision
sentencepiece
torch
google-generativeai
gradio
pillow
pytesseract
requests
dotenv